move util functions outside Searcher
