+++ server/api/search/SearchCahce.js
@@ -4,79 +4,42 @@ import JSONStream from 'JSONStream';
 import Writer from './Streamables';
 import createPrefixModel from './model/Prefix';
 
-const validateInputIsArray = (input, funcName) => {
-  if (!Array.isArray(input)) {
-    throw new TypeError(`The argument to ${funcName} must be an array`);
-  }
-};
-
-// remove instances of more than one whitespace
-const normalizeCompletion = (completionMaxChars, originalCompletions) =>
-  originalCompletions
-    .slice(0, completionMaxChars)
-    .toLowerCase()
-    .trim()
-    .replace(/\s{2,}/g, ' ');
-
-// remove instances of more than one whitespace
-const normalizePrefix = (prefixMaxChars, originalQuery) =>
-  originalQuery
-    .slice(0, prefixMaxChars)
-    .toLowerCase()
-    .replace(/\s{2,}/g, ' ');
-
-const extractPrefixes = (prefixMaxChars, prefixMinChars, completion) => {
-  const normalized = normalizePrefix(prefixMaxChars, completion);
-
-  const prefixes = [];
-  for (let i = prefixMinChars; i <= normalized.length; i++) {
-    prefixes.push(normalized.slice(0, i));
-  }
-  return prefixes;
-};
-
-const toFullPrefix = (prefix, token) => `${token}:${prefix}`;
-
-const DefaultConfig = {
-  maxMemory: 500,
-  suggestionCount: 10,
-  prefixMinChars: 1,
-  prefixMaxChars: 15,
-  completionMaxChars: 50,
-  bucketLimit: 50,
-};
-
-const mongoPersist = async (redisClient, prefix, token) => {
-  const prefixWithTenant = toFullPrefix(prefix, token);
-  const completions = await redisClient.zrangeAsync(
-    prefixWithTenant,
-    0,
-    -1,
-    'WITHSCORES'
-  );
-
-  const prefixModel = createPrefixModel(token);
-  if (completions.length === 0) {
-    await prefixModel.findOneAndDelete({ prefix });
-  } else {
-    await prefixModel.createIndexes({ prefix: 'text', background: true });
-    await prefixModel.findOneAndUpdate(
-      { prefix },
-      { $set: { completions } },
-      { new: true, upsert: true }
-    );
-  }
-};
+import {
+  validateInputIsArray,
+  normalizeCompletion,
+  normalizePrefix,
+  extractPrefixes,
+  toFullPrefix,
+} from './utils';
 
 const persistPrefixes = async (redisClient, prefixes, token) => {
   for (var i = 0; i < prefixes.length; i++) {
-    await mongoPersist(redisClient, prefixes[i], token);
+    const prefixWithTenant = toFullPrefix(prefix, token);
+    const completions = await redisClient.zrangeAsync(
+      prefixWithTenant,
+      0,
+      -1,
+      'WITHSCORES'
+    );
+
+    const prefixModel = createPrefixModel(token);
+    if (completions.length === 0) {
+      await prefixModel.findOneAndDelete({ prefix });
+    } else {
+      await prefixModel.createIndexes({ prefix: 'text', background: true });
+      await prefixModel.findOneAndUpdate(
+        { prefix },
+        { $set: { completions } },
+        { new: true, upsert: true }
+      );
+    }
   }
 };
-const mongoLoad = async (redisClient, prefix, token) => {
+const syncRedisWithMongo = async (redisClient, prefix, token) => {
   const prefixModel = createPrefixModel(token);
   const singleModel = await prefixModel.findOne({ prefix });
-  const completions = singleModel && singleModel.completions ? singleModel.completions : [];
+  const completions =
+    singleModel && singleModel.completions ? singleModel.completions : [];
 
   const prefixWithTenant = toFullPrefix(prefix, token);
 
@@ -102,7 +65,7 @@ const getCompletionsCount = async (
   const count = await redisClient.zcountAsync(prefixWithTenant, '-inf', '+inf');
 
   if (count === 0) {
-    await mongoLoad(redisClient, prefix, tenant);
+    await syncRedisWithMongo(redisClient, prefix, tenant);
     return await redisClient.zcountAsync(prefixWithTenant, '-inf', '+inf');
   } else {
     return count;
@@ -177,10 +140,15 @@ const apiSearch = instance => async (prefixQuery, token, opts = {}) => {
   const prefix = normalizePrefix(prefixMaxChars, prefixQuery);
   const prefixWithTenant = toFullPrefix(prefix, token);
 
-  const result = await redisClient.zrangeAsync(prefixWithTenant, 0, suggestionCount - 1, 'WITHSCORES');
+  const result = await redisClient.zrangeAsync(
+    prefixWithTenant,
+    0,
+    suggestionCount - 1,
+    'WITHSCORES'
+  );
 
   if (result.length === 0) {
-    await mongoLoad(redisClient, prefix, token);
+    await syncRedisWithMongo(redisClient, prefix, token);
     return await redisClient.zrangeAsync(...args);
   } else {
     return result;
@@ -271,6 +239,15 @@ const dumpFile = instance => (filePath, token) => {
     json.pipe(parser).pipe(writer);
     writer.on('finish', () => resolve('Import finished'));
   });
+};
+
+const DefaultConfig = {
+  maxMemory: 500,
+  suggestionCount: 10,
+  prefixMinChars: 1,
+  prefixMaxChars: 15,
+  completionMaxChars: 50,
+  bucketLimit: 50,
 };
 
 const init = (redisClient, config = DefaultConfig) => {
